@pcrespov review: move db mock fixtures in pytest-simcore
