+++ services/director-v2/tests/integration/conftest.py
@@ -7,43 +7,6 @@ from faker import Faker
 from models_library.projects import ProjectAtDB, ProjectType
 from simcore_postgres_database.models.comp_tasks import comp_tasks
 from simcore_postgres_database.models.projects import projects
-from simcore_postgres_database.models.users import UserRole, UserStatus, users
-
-
-@pytest.fixture()
-def registered_user(
-    postgres_db: sa.engine.Engine, faker: Faker
-) -> Iterator[Callable[..., Dict]]:
-    created_user_ids = []
-
-    def creator(**user_kwargs) -> Dict[str, Any]:
-        with postgres_db.connect() as con:
-            # removes all users before continuing
-            user_config = {
-                "id": len(created_user_ids) + 1,
-                "name": faker.name(),
-                "email": faker.email(),
-                "password_hash": faker.password(),
-                "status": UserStatus.ACTIVE,
-                "role": UserRole.USER,
-            }
-            user_config.update(user_kwargs)
-
-            con.execute(
-                users.insert().values(user_config).returning(sa.literal_column("*"))
-            )
-            # this is needed to get the primary_gid correctly
-            result = con.execute(
-                sa.select([users]).where(users.c.id == user_config["id"])
-            )
-            user = result.first()
-            created_user_ids.append(user["id"])
-        return dict(user)
-
-    yield creator
-
-    with postgres_db.connect() as con:
-        con.execute(users.delete().where(users.c.id.in_(created_user_ids)))
 
 
 @pytest.fixture
